+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add head_js block into base html template
 - list teacher's organizations in the teacher dashboard sidebar.
+- Added Certificates in the dashboard
 
 ### Fixed